attendence/requirements.txt
streamlit
opencv-python
numpy
pandas
Pillow
matplotlib
scikit-learn
